re-fixed the calc_12C bug and generate_exp_guess
The thing is: calc_12C wasn't the problem, it was that I was separating
by sample - which is a single bottle that goes into the LC/MS, thus a
single time point.

MAYBE MORE IMPORTANTLY I updated the error checking on
generate_exp_guess so that it deals with zeros better (i.e., doesn't
fail when you try to calculate log(0).